cod system
The project that I am creating is a database for a cash on delivery system. It is 
related to the post office. This is done by sellers selling a product online or by 
paying only one delivery charge. After the seller has delivered any ordered 
product to the post office, the product will be taken to the relevant address through 
the postman including all the details of the recipient, where payment should be 
made in cash after receiving the parcel and the postman should give the amount 
to the post office.